* Set correct length for samples being dragged in
* Create TimeRuler

* Add Length to parts

+++ TrackerDAW/Song/Part.cs
@@ -14,6 +14,6 @@ namespace TrackerDAW
         [JsonProperty] public double Offset { get; set; }
 
         public abstract Part Clone();
-        public abstract float GetLength();
+        public abstract double GetLength();
     }
 }
